adding additional apis and documents

+++ src/main/java/com/aconex/viewer/model/Project.java
@@ -20,7 +20,7 @@ import java.util.Set;
 public class Project implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long projectId;
+    private int projectId;
 
     @NotEmpty
     @Column
@@ -39,11 +39,11 @@ public class Project implements Serializable {
     @OneToMany(mappedBy = "project", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     private Set<Floor> floors;
 
-    public Long getProjectId() {
+    public int getProjectId() {
         return projectId;
     }
 
-    public void setProjectId(Long projectId) {
+    public void setProjectId(int projectId) {
         this.projectId = projectId;
     }
